+++ CHANGES.rst
@@ -48,6 +48,10 @@ Changelog
 
 - Fix rebuild_i18n script [cekk]
 
+- Replaced "getLastCommentUrl" index/metadata with "getLastCommentId".
+  This fixes a bug when we have frontend url different from backend.
+  [cekk]
+
 3.4 (2013-05-11)
 ----------------